Word movers distance in search and refactor
Auth level to anonymous

Access file after download only

Word2vec download only when needed

Resource config

Add wheel for package

Lemmatization disabled in word mover distance

Using automatic garbage collection of python

Remove imports of gc

Remove usage of search service from source watcher (#336)

Remove usage of search service from source watcher

Remove usage of search service from source watcher
