+++ src/Diagnostics.AIProjects/SearchAPI/AppConfig/appconfig.json
@@ -1,4 +1,5 @@
 {
+    "MODEL_SYNC_ENABLED": true,
     "APP_ID": "__APP_ID__",
     "WHITELISTED_APPS": "__WHITELISTED_APPS_",
     "STORAGE_ACCOUNT_NAME": "__STORAGE_ACCOUNT_NAME__",
