+++ src/Diagnostics.AIProjects/SearchAPI/SearchModule/MessageStrings.py
@@ -1,0 +1,3 @@
+loadModelMessage = "TextSearchModule: Loading model for product {0}: "
+refreshModelMessage = "TextSearchModule: Refresh model request for product {0}: "
+fileMissingMessage = "One or more of model file(s) are missing"
