+++ src/Diagnostics.AIProjects/SearchAPI/requirements.txt
@@ -9,3 +9,4 @@ azure-storage==0.36.0
 pytz==2018.9
 pywin32
 googletrans==2.4.0
+pyemd-0.5.1-cp36-cp36m-win_amd64.whl
