+++ src/Diagnostics.AIProjects/TrainingAPI/__app__/TrainingModule/Exceptions.py
@@ -1,0 +1,8 @@
+class TrainingException(Exception):
+    pass
+
+class PublishingException(Exception):
+    pass
+
+class ResourceConfigDownloadFailed(Exception):
+    pass
