+++ src/Diagnostics.AIProjects/TrainingAPI/__app__/TrainingModule/Utilities.py
@@ -1,0 +1,5 @@
+import os
+
+def cleanFolder(folderPath):
+    for f in os.listdir(folderPath):
+        os.remove(os.path.join(folderPath, f))
