+++ src/Diagnostics.AIProjects/TrainingAPI/__app__/TriggerTraining/__init__.py
@@ -1,4 +1,5 @@
-import logging, asyncio, json
+import logging, asyncio, json, os
+from pathlib import Path
 import azure.functions as func
 from __app__.AppSettings.AppSettings import appSettings
 from __app__.TrainingModule.HandleRequest import triggerTrainingMethod
@@ -18,6 +19,9 @@ async def main(req: func.HttpRequest) -> func.HttpResponse:
         if "trainingConfig" in req_data:
             trainingConfig = TrainingConfig(json.loads(req_data["trainingConfig"]))
             sah = StorageAccountHelper()
+            if trainingConfig.modelType == "WmdSearchModel":
+                if not Path(os.path.join(appSettings.WORD2VEC_PATH, appSettings.WORD2VEC_MODEL_NAME)).exists():
+                    await sah.downloadFile("word2vec/w2vModel.bin", appSettings.WORD2VEC_PATH)
             await sah.downloadFile("resourceConfig/config.json")
             await sah.downloadFile("{0}/testCases.json".format(productId), "TestingModule")
             await sah.downloadFile("{0}/rawdata/SampleUtterances.json".format(productId), "rawdata_{0}".format(productId))
