+++ src/Diagnostics.AIProjects/TrainingAPI/__app__/TriggerTraining/function.json
@@ -2,7 +2,7 @@
   "scriptFile": "__init__.py",
   "bindings": [
     {
-      "authLevel": "function",
+      "authLevel": "anonymous",
       "type": "httpTrigger",
       "direction": "in",
       "name": "req",
